requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.3.0
cachetools==4.0.0
chardet==3.0.4
discord.py==1.2.5
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
httplib2==0.14.0
idna==2.8
idna-ssl==1.1.0
multidict==4.7.0
nltk==3.4.5
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
rsa==4.0
six==1.13.0
typing-extensions==3.7.4.1
uritemplate==3.0.0
websockets==6.0
wn==0.0.22
yarl==1.4.2
